added the rebuild option to build.sh

+++ build.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
-rm -rf build
-mkdir build
+if [ "$1" = "rebuild" ] || [ ! -d "build/" ]; then
+    rm -rf build
+    mkdir build
+
+    (
+      cd build || exit 1
+      cmake -DCMAKE_TOOLCHAIN_FILE=../toolchain.cmake -DCMAKE_BUILD_TYPE=Debug ..
+    )
+fi
 
 (
   cd build || exit 1
-  cmake -DCMAKE_TOOLCHAIN_FILE=../toolchain.cmake ..
   make -j "$(nproc)"
-
   cp modules/cli/cli cli
 )
